Update IgnoreMe.py
now accomodates for number of shifts by making sure that nobody has more than average. There are 78 shifts to fill in 10 weeks and the average is just 78 / number of ras. Nmber of ras happens to be length of the temporary dictonary i make here... 

TODO
make it read the actual python dictionary not the temp one in file